Completely generic main, rest implemented by hooks
