+++ protocol/pjrc.mk
@@ -1,6 +1,6 @@
 PJRC_DIR = protocol/pjrc
 
-SRC +=	$(PJRC_DIR)/main.c \
+SRC +=	$(PJRC_DIR)/pjrchooks.c \
 	$(PJRC_DIR)/pjrc.c \
 	$(PJRC_DIR)/usb_keyboard.c \
 	$(PJRC_DIR)/usb_debug.c \
