+++ protocol/vusb/vusb.h
@@ -21,7 +21,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "host_driver.h"
 
 
-host_driver_t *vusb_driver(void);
+extern host_driver_t vusb_driver;
 void vusb_transfer_keyboard(void);
 
 #endif
